@@ -11,7 +11,7 @@ classifiers =
 
 [options]
 zip_safe = False
-python_requires = >=3.8
+python_requires = >=3.9
 include_package_data = True
 packages = find:
 install_requires =
